feat(navigation.component.ts): improve navigation component UI and add link icon and arrow
The changes in this commit improve the UI of the navigation component by adding an icon and arrow to each link. The link structure has been modified to include a link group that contains the link text and icon. The link arrow has also been added to indicate that the link is clickable. Additionally, some styling changes have been made to the navigation component, including setting the height and margin, adding a border to each list item, and removing the border from the last list item.